+++ src/components/Title.jsx
@@ -1,22 +1,22 @@
 import { 
     useLocation, 
-    useEffect 
+    useEffect,
+    Outlet
 } from './barrel_module/Barrel.jsx';
 
-function Title({ children }) {
+function Title() {
   const location = useLocation();
 
   useEffect(() => {
     const titles = {
       "/": "Beranda | CampusHub",
       "/dashboard": "Dashboard | CampusHub",
-      "/gallery": "Galeri | CampusHub",
     };
 
     document.title = titles[location.pathname] || "CampusHub";
   }, [location]);
 
-  return <div>{children}</div>;
+  return <Outlet />;
 }
 
 export default Title;